Add YT audio playing
This took a lot of fiddling to get right, since (during my first write of this) an end event could come from the stream ending or from someone requesting something new

It's now split. If a queue already exists, just add on to it. If it doesn't, create a queue and loop over it